requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r ./base.txt

gunicorn==21.2.0
psycopg2==2.9.9
Collectfast==2.2.0

# Django
# ------------------------------------------------------------------------------
django-storages[boto3]==1.14.2
django-anymail[mailgun]==10.1
